moved get_city_id to helper file
