+++ app/helper.py
@@ -144,3 +144,18 @@ def load_data_rent_visual():
   df_mew['date']= pd.to_datetime(df_mew['date'])
   return df_melt, df_me, df_mew
 
+def get_city_id(city_name, state_abbreviation):
+    load_dotenv()
+    database_url = os.getenv('PRODUCTION_DATABASE_URL')
+    engine = sqlalchemy.create_engine(database_url)
+    city_name = city_name.title()
+    state_abbreviation = state_abbreviation.upper()
+    query = f'''
+                 SELECT Cities.city_id
+                FROM CITIES
+                LEFT JOIN STATES ON CITIES.state_id=STATES.state_id
+                Where cities.city_name = \'{city_name}\' and states.state_abbreviation = \'{state_abbreviation}\'
+                '''
+    query_result = engine.execute(query)
+    my_return = [each[0] for each in query_result]
+    return my_return[0]
